Pre-allocate request buffers in the request structure
Previously, if a request yielded from a coroutine, its buffer (and hence all
the request information) would be lost the moment the that thread got
another request to work on.

This uses a lot of memory (4kb more for each request), but is preallocated
on program start so there isn't a runtime malloc/free overhead.