Compress cached files
This uses zlib to deflate the contents. Right now, the compressed contents
will be served unless there was an error during the compression, in which
case, the uncompressed version will be then served. No care is taken to
see if the client accepts compressed responses; this must be fixed before
merging on main.

In addition to the TODO items,  error handling in _compress_cached_file()
should be improved.