Don't give up so easily when file descriptors are scarce
This will try to open a file a number of times -- decrementing a counter
every time the process file descriptor table is full, and yielding the
coroutine at that point.  This should, in theory, give some time to free up
a file descriptor.  If that counter goes to zero, we just give up and return
a 503 (Service Unavailable) error.  All other errors (Forbidden, Not Found,
etc) as handled as usual.