Refactor request parsing to reduce lwan_request_t structure
This has been accomplished by splitting the request structure into two:
one that is kept for the whole duration of the program, and another that
is used only temporarily while parsing the request.

On x86-64, this means we went from 280bytes to 184bytes. There's still
some work left that might reduce this even more.