Allocate coroutine string buffer on stack
This further reduces the amount of dynamic memory allocation in the
fast path.

The important bit in this change is that the process_request_coro()
function should never return: coro_yield() should be called to
terminate it.  This way, the string buffer remains alive when the
deferred callbacks are executed to clean it up.