Align coroutine stack on x86-64
Unaligned pointers shouldn't cause issues on x86-64 -- except when SIMD
is used.  Since the stack is assumed to be aligned by gcc, it ends up
using SIMD instructions at will -- which might cause problems if it's
not actually aligned.  Since the coroutine stuff messes with the stack
pointer, it was a miracle that this thing worked for so long.

Found with the help of address sanitizer.