Before calling chroot(), ensure no directories are open
This works only on Linux, and even then, it is racy and borderline
useless from a security perspective.

It is more useful, however, in the cases where people add the
"straitjack" section in the configuration file after declaring an
instance of the file serving module, and the directory is outside the
chroot.