Implement shorthand syntax for handlers in config files
This turns sections like this:

	prefix /some/prefix {
		handler = hello_world
	}

Into this:

	&hello_world /some/prefix {
	}

The '&' was chosen since it's the address operator in C, and this
is essentially obtaining the address of the "hello_world" symbol.

It's possible to implement the same feature without requiring an
extra symbol, however, a module may have the same name, in which
case things could fail silently.