Bundle auto-index icons in executable
With this change, icons for folders, files, and parent directory
are bundled in the executable.  Because of this, it's possible to
have automatic directory indexes without having to have a /icons
endpoint serving files provided by Lwan.

These icons can be changed in two ways: (1) by modifying the icons
themselves inside wwwroot/icons and rebuilding Lwan and (2) by
changing the auto index template.  If (2) is chosen, user will
have to serve their own icons as somewhere.