Reorganize source directory and clean up CMakeLists.txt files
Source code has been moved to a "src"
directory; a "lib" subdirectory contains the
source code for the "liblwan.a" library, and   ├── src
a "bin" subdirectory contains code for the     │   ├── bin
Lwan server, the test runner, and tools used   │   │   ├── lwan
during the build process (i.e.  mimegen and    │   │   ├── testrunner
bin2hex).                                      │   │   └── tools
                                               │   ├── cmake
A "cmake" subdirectory contains CMake-related  │   ├── lib
files, such as additional macros and files     │   │   └── missing
that will be processed by CMake, such as the   │   │       └── sys
template for the pkg-config file.              │   ├── samples
                                               │   │   ├── freegeoip
The "samples" subdirectory contains the        │   │   └── techempower
FreeGeoIP and TechEmpower samples.             │   └── scripts
                                               └── wwwroot
The "scripts" subdirectory contains the test       └── icons
suite and benchmarking scripts.

The "wwwroot" directory in the project root remains as it was before.